==> Crickbuzz Player Scraper/Detailed/3. Merging the Dataset.ipynb <==
import pandas as pd

df = pd.read_csv('players_record.csv')
df.head()

### 1. Defining Functions to get Column Names and Data

def get_cols(file_path, sep):  
    cols = []
    col_names = pd.read_csv(file_path).columns
    for _ in ['Test','ODI','T20I','IPL']:
        for col_name in col_names[1:]:
            cols.append(sep + "_" + _ + "_" + col_name)
    return cols

def get_data(file_path, seq_len):
    
    df_ = pd.read_csv(file_path)

    test = ['-' for i in range(seq_len)]
    odi  = ['-' for i in range(seq_len)]
    t20  = ['-' for i in range(seq_len)]
    ipl  = ['-' for i in range(seq_len)]

    for i in df_.values:   
        if i[0] == 'Test':
            test = i[1:]
        if i[0] == 'ODI':
            odi = i[1:]
        if i[0] == 'T20I':
            t20 = i[1:]
        if i[0] == 'IPL':
            ipl = i[1:]

        final =  [_ for _ in test]
        final += [_ for _ in odi ]
        final += [_ for _ in t20 ]
        final += [_ for _ in ipl ]

    return final

### 2. Finding the Balling and Batting Data

bat_data = []
bow_data = []


for i in df.values:
    bat_path = i[-2]
    bat_data.append(get_data(bat_path,13))
    
    bow_path = i[-1]
    bow_data.append(get_data(bow_path,12))

### 3. Creating the DataFrame for Balling and Batting Data

df_bat = pd.DataFrame(bat_data, columns = get_cols('Scores/bat_1413.csv', 'BT'))
df_bat.to_csv('bat.csv', index = False)

df_bow = pd.DataFrame(bow_data, columns = get_cols('Scores/bow_1413.csv', 'BW'))
df_bow.to_csv('ball.csv', index = False)

### 4. Combining and Exporting the DataFrame

df_fin = pd.concat((df,df_bat,df_bow), axis = 1)
df_fin.to_csv('final.csv', index = False)
