==> Crickbuzz Player Scraper/Detailed/4. Column Documentation.ipynb <==
import json
import requests
import numpy as np
import pandas as pd
from bs4 import BeautifulSoup

link = pd.read_csv('cricket.csv')['link'].values[0]

soup = BeautifulSoup(requests.get(link).content,'lxml')

### 1.Scraping the Column Data

batt_col = [[i.text,i.get('title')] for i in soup.find_all('table')[0].find_all('th')][1:]
ball_col = [[i.text,i.get('title')] for i in soup.find_all('table')[1].find_all('th')][1:]

### 2. Adding Documentation to each Column

dct = {}

for i in batt_col:
    for j in ['Test','ODI','T20I','IPL']:        
        key = 'BT_' + j + '_' + i[0]
        value = j + ' ' + i[1] + ' ' + 'as a battsman'
        dct[key] = value
        
        
for i in ball_col:
    for j in ['Test','ODI','T20I','IPL']:        
        key = 'BT_' + j + '_' + i[0]
        value = j + ' ' + i[1] + ' ' + 'as a baller'
        dct[key] = value


### 3. Saving Columns Manually

for i in  ['name','team',
        'date_of_birth','palce_of_birth',
        'role',    'height',
        'test_bat_rank','odi_bat_rank','t20_bat_rank',
        'test_bow_rank','odi_bow_rank',
        't20_bow_rank','link', 'bat_path','bow_path']:

    dct[i] = i + ' of the player'

### 4. Saving it as a JSON

fd = open('columns.json','w')
fd.write(json.dumps(dct))
fd.close()
